Refactor portals implementation in separate service part
